requirements.txt
ir-measures
flake8
autopep8
python-dotenv
openai
pandas
rank_bm25
git+https://github.com/castorini/pygaggle.git
huggingface-hub
faiss-gpu
protobuf==3.20.*
beautifulsoup4
lxml
tenacity
peft